Update unlock.py
- implemented the "store progress" by taskula
- IMEI check and warning
- minor speed improvement
- auto stop when reach 17 digits
- tell if it could find the code

Use it in RAM drive or else it can wear out your HDD / SSD (so many write cycle)!

+++ unlock.py
@@ -7,30 +7,38 @@ https://en.wikipedia.org/wiki/Luhn_algorithm
 """
 
 import time
-#from flashbootlib import test
 import os
 import math
 
+import configparser
+from os import path
+
 ##########################################################################################################################
 
-def tryUnlockBootloader(checksum):
+def tryUnlockBootloader():
 
     unlock      = False
-    algoOEMcode = 1000000000000000 #base
-    
-    while(unlock == False):
+
+    increment = int(luhn_checksum(imei)+math.sqrt(imei)*1024)
+    algoOEMcode = int(progress.get(str(imei), 'last_attempt', fallback=1000000000000000)) #base
+    if not str(imei) in progress:
+        progress[str(imei)] = {}
+    progress[str(imei)]['last_attempt'] = str(algoOEMcode)
+    # Load possible previous progress
+    progress.read(progress_file)
+
+    while((unlock == False) and (algoOEMcode < 10000000000000000)):
         sdrout = str(os.system('fastboot oem unlock '+str(algoOEMcode)))
         sdrout = sdrout.split(' ')
         for i in sdrout:
             if i == 'success':
+                unlock = True
                 return(algoOEMcode)
-
-        algoOEMcode = algoIncrementChecksum(algoOEMcode, checksum)
-
-
-def algoIncrementChecksum(genOEMcode, checksum):
-    genOEMcode+=int(checksum+math.sqrt(imei)*1024)
-    return(genOEMcode)
+        algoOEMcode+=increment
+        progress.set(str(imei), 'last_attempt', str(algoOEMcode))
+        with open(progress_file, 'w') as f:
+            progress.write(f)
+    return(-1)
 
 def luhn_checksum(imei):
     def digits_of(n):
@@ -45,6 +53,12 @@ def luhn_checksum(imei):
     return checksum % 10
 
 ##########################################################################################################################
+progress_file = 'progress.ini'
+progress = configparser.ConfigParser()
+if not path.exists(progress_file):
+    with open(progress_file, 'w') as f:
+        progress.write(f)
+progress.read(progress_file)
 
 print('\n\n           Unlock Bootloader script - By SkyEmie_\'')
 print('\n\n  (Please enable USB DEBBUG and OEM UNLOCK if the device isn\'t appear..)')
@@ -53,18 +67,34 @@ input(' Press any key to detect device..\n')
 
 os.system('adb devices')
 
-imei     = int(input('Type IMEI digit :'))
-checksum = luhn_checksum(imei)
-input('Press any key to reboot your device..\n')
-os.system('adb reboot bootloader')
-input('Press any key when your device is ready.. (This may take time, depending on your cpu/serial port)\n')
 
-codeOEM = tryUnlockBootloader(checksum)
+if progress.sections():
+    print('The following IMEIs were already saved')
+    print(*progress.sections(), sep = ', ')
 
-os.system('fastboot getvar unlocked')
-os.system('fastboot reboot')
+def start_bruteforce():
+    input('Press any key to reboot your device..\n')
+    os.system('adb reboot bootloader')
+    input('Press any key when your device is ready.. (This may take time, depending on your cpu/serial port)\n')
+    codeOEM = tryUnlockBootloader()
+    if codeOEM > 0:
+        os.system('fastboot getvar unlocked')
+        os.system('fastboot reboot')
+        print('\n\nDevice unlock ! OEM CODE : '+codeOEM)
+        print('(Keep it safe)\n')
+    else:
+        print('Could not find the OEM CODE')
 
-print('\n\nDevice unlock ! OEM CODE : '+codeOEM)
-print('(Keep it safe)\n')
+imei = int(input('Type IMEI digit :'))
+start = 'no'
+if luhn_checksum(imei) == 0:
+    start = 'yes'
+else:
+    print('IMEI number seems invalid.')
+    start = input('Would you like to try anyway? (type \'yes\' or \'no\')')
+
+if start == 'yes':
+    start_bruteforce()
+
 input('Press any key to exit..\n')
 exit()
